feat: Responsive design and images, html markup changes, color changes
Restaurant home was made responsive with use of flexbox and css grid.
Images were also made responsive accordingly using min-height of low value
and
width as a percentage. The min-height was given in such a way so as to
achieve natural aspect ratio.
Colors were improved to bring better UX
and achieve minimum of 4.5:1 contrast.
html markup was made utf-8 encoding and meta viewport markup was
also added.
FInally to some code spacing changes were done to remove white spaces
and to align the code properly.